requirements.txt
Flask
Werkzeug
python-dotenv
google-generativeai
endee
pandas
beautifulsoup4
requests
pypdf
openpyxl
tabulate
gunicorn
